+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Changed
-- Instead of reading types by re-parsing downloaded SHACL files, the GraphQL Schema file is now used as input.
 
+## 1.0.0-beta.7 - 2023-06-20
 ### Added
 - Test cases for GraphQLSchema class read from a GQL Schema file.
+
+### Changed
+- Instead of reading types by re-parsing downloaded SHACL files, the GraphQL Schema file is now used as input.
 
 ### Removed
 - Logger library typescript-logging removed, since it was not being used yet.